removido arquivos de services desnecessarios
