--- src/Lumen/Services/BaseService.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace Bludata\Lumen\Services;
-
-use Bludata\Lumen\Common\Interfaces\BaseServiceInterface;
-
-abstract class BaseService implements BaseServiceInterface
-{
-    /**
-     * @var Bludata\Lumen\Repositories\BaseRepository
-     */
-    protected $mainRepository;
-
-    public function getMainRepository()
-    {
-        return $this->mainRepository;
-    }
-}
