--- src/Lumen/Traits/Services/ReadTrait.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Bludata\Lumen\Traits\Services;
-
-trait ReadTrait
-{
-    public function findAll(array $filters = null)
-    {
-        return $this->mainRepository->findAll()->withFilters($filters);
-    }
-
-    public function find($id)
-    {
-        return $this->mainRepository->find($id);
-    }
-}
