--- src/Lumen/Traits/Tests/Services/DeleteTrait.php
@@ -1,22 +1,0 @@
-<?php
-
-namespace Bludata\Lumen\Traits\Tests\Services;
-
-use Bludata\Doctrine\Common\Interfaces\BaseEntityInterface;
-
-trait DeleteTrait
-{
-    public function testRemove(BaseEntityInterface $entity = null)
-    {
-        $entity = $entity
-                    ? $entity
-                    : $this->getRepositoryTest()->getFlushedMockObject();
-
-        $entityRemoved = $this->getService()
-                ->remove($entity->getId())
-                ->flush();
-
-        $this->assertInstanceOf('DateTime', $entityRemoved->getDeletedAt());
-        $this->assertNull($this->getService()->getMainRepository()->find($entity->getId(), false));
-    }
-}
